pyproject.toml
[tool.poetry]
name = "sreality-scraper"
version = "0.1.0"
description = ""
authors = ["Martin Dluhoš <dluhos.martin@gmail.com>"]
readme = "README.md"
packages = [{include = "sreality_scraper"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
